feat:add landing page and route
